+++ lib/src/enums/thumbnail_quality.dart
@@ -5,17 +5,17 @@
 /// Quality of YouTube video thumbnail.
 class ThumbnailQuality {
   /// 120*90
-  static const String defaultQuality = 'default.webp';
+  static const String defaultQuality = 'default';
 
   /// 320*180
-  static const String medium = 'mqdefault.webp';
+  static const String medium = 'mqdefault';
 
   /// 480*360
-  static const String high = 'hqdefault.webp';
+  static const String high = 'hqdefault';
 
   /// 640*480
-  static const String standard = 'sddefault.webp';
+  static const String standard = 'sddefault';
 
   /// Unscaled thumbnail
-  static const String max = 'maxresdefault.webp';
+  static const String max = 'maxresdefault';
 }