+++ lib/src/widgets/youtube_player_builder.dart
@@ -11,11 +11,19 @@ class YoutubePlayerBuilder extends StatefulWidget {
   /// Builds the widget below this [builder].
   final Widget Function(BuildContext, Widget) builder;
 
+  /// Callback to notify that the player has entered fullscreen.
+  final VoidCallback onEnterFullScreen;
+
+  /// Callback to notify that the player has exited fullscreen.
+  final VoidCallback onExitFullScreen;
+
   /// Builder for [YoutubePlayer] that supports switching between fullscreen and normal mode.
   const YoutubePlayerBuilder({
     Key key,
     @required this.player,
     @required this.builder,
+    this.onEnterFullScreen,
+    this.onExitFullScreen,
   }) : super(key: key);
 
   @override
@@ -45,9 +53,11 @@ class _YoutubePlayerBuilderState extends State<YoutubePlayerBuilder>
     if (physicalSize.width > physicalSize.height) {
       controller.updateValue(controller.value.copyWith(isFullScreen: true));
       SystemChrome.setEnabledSystemUIOverlays([]);
+      if (widget.onEnterFullScreen != null) widget.onEnterFullScreen();
     } else {
       controller.updateValue(controller.value.copyWith(isFullScreen: false));
       SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
+      if (widget.onExitFullScreen != null) widget.onEnterFullScreen();
     }
     super.didChangeMetrics();
   }
@@ -56,11 +66,22 @@ class _YoutubePlayerBuilderState extends State<YoutubePlayerBuilder>
   Widget build(BuildContext context) {
     final _player = Container(
       key: playerKey,
-      child: widget.player,
+      child: WillPopScope(
+        onWillPop: () async {
+          final controller = widget.player.controller;
+          if (controller.value.isFullScreen) {
+            widget.player.controller.toggleFullScreenMode();
+            return false;
+          }
+          return true;
+        },
+        child: widget.player,
+      ),
     );
     final child = widget.builder(context, _player);
     return OrientationBuilder(
-        builder: (context, orientation) =>
-            orientation == Orientation.portrait ? child : _player);
+      builder: (context, orientation) =>
+          orientation == Orientation.portrait ? child : _player,
+    );
   }
 }