+++ test/youtube_player_flutter_test.dart
@@ -23,7 +23,7 @@ Widget buildPlayer({
   void Function() onReady,
   ProgressBarColors progressBarColors,
   Color progressIndicatorColor,
-  String thumbnailUrl,
+  Widget thumbnail,
   EdgeInsetsGeometry actionsPadding = const EdgeInsets.all(8.0),
 }) {
   return TestApp(
@@ -40,7 +40,7 @@ Widget buildPlayer({
       onReady: onReady,
       progressColors: progressBarColors,
       progressIndicatorColor: progressIndicatorColor,
-      thumbnailUrl: thumbnailUrl,
+      thumbnail: thumbnail,
       topActions: topActions,
     ),
   );